py-version/requirements.txt
openai-agents
python-dotenv
httpx[socks]
